fix js type checks and jsre
