+++ .gitignore
@@ -1,4 +1,3 @@
 /subprojects/*
 !/subprojects/packagefiles
 !/subprojects/*.wrap
-.cache/
